+++ packages/core/pluginutils.d.ts
@@ -1,14 +1,14 @@
 /**
  * Return an 8 character hash safe to use in urls and ids
  */
-export function hash(str: string): string
+export declare function hash(str: string): string
 
 /**
  * Dedent a string to look pretty
  */
-export function dedent(str: string): string
+export declare function dedent(str: string): string
 
 /**
  * Escape a string to be used as an attribute value
  */
-export function escapeAttr(value: string): string
+export declare function escapeAttr(value: string): string